Clean up compiler warnings

+++ java/src/jmri/jmrix/dccpp/DCCppReply.java
@@ -337,7 +337,7 @@ public class DCCppReply extends jmri.jmrix.AbstractMRReply {
 	if (this.isProgramReply()) {
 	    Matcher m = match(this.toString(), DCCppConstants.PROGRAM_REPLY_REGEX, "ProgramReply");
 	    if (m != null) {
-		if (m.group(2).equals(Integer.toString((int)DCCppConstants.PROG_WRITE_CV_BIT)))
+		if (m.group(2).equals(Integer.toString(DCCppConstants.PROG_WRITE_CV_BIT)))
 		    return(m.group(6));
 		else
 		    return(m.group(4));
